backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
djangorestframework==3.16.0
idna==3.10
joblib==1.5.0
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.4.0
